Ci checks passing

+++ codec/src/delta.rs
@@ -540,11 +540,7 @@ fn encode_delta_payload_from_updates(
             limits,
             |writer| {
                 encode_update_body_sparse_packed_from_updates(
-                    schema,
-                    updates,
-                    limits,
-                    &lookup,
-                    writer,
+                    schema, updates, limits, &lookup, writer,
                 )
             },
         )?;
@@ -1051,7 +1047,7 @@ fn build_component_lookup(schema: &schema::Schema) -> ComponentLookup {
         let id = component.id.get() as usize;
         index[id] = Some(component_index);
         let bits = required_bits(component.fields.len().saturating_sub(1) as u64);
-        index_bits[id] = bits as u8;
+        index_bits[id] = bits;
     }
     ComponentLookup { index, index_bits }
 }
